allow fonts made out of lists of font files
A font is now a list of font faces, when a glyph is required it is taken
from the first face in the list that contains that glyph. This allows
support for fonts like google noto that consists of many files all
together representing the complete font

Additionally the font cache has changed, it now never deletes fonts
loaded, you have to do so on your own by calling a function.