optimize blending
this results in 10-30% speedup for the example3 usage patterns
it also fixes a problem with low bit textures (e.g. 16 Bit per pixel)
where the gamma blending can lead to strange rectangular patterns
visible in the area of the texture but there no glyph data actually was.
This is due to the rounding errors introduced for the gamma corrections
and the shifting of the data when reading and writing pixels